travis: Use Xcode 9.2 on macOS
Xcode 8.3, to which there recently was a switch, spits out a warning for
the access of the potentially unaligned access to ip6_plen in ip-packet.c,
which we explicitly read via untoh16() hence the access to that pointer
is not actually unaligned. Xcode 9 does not report this false positive,
so let's use that for now (until Xcode 9+ is the default).